requirements.txt
flask
python-dotenv
reportlab
huggingface_hub
